requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
cloudinary==1.36.0
coverage==7.6.10
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.17
django-allauth==0.57.2
execnet==2.1.1
gunicorn==20.1.0
iniconfig==2.0.0
oauthlib==3.2.2
pathspec==0.12.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
pytest-xdist==3.6.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
requests>=2.28.0
sqlparse==0.5.3
urllib3==1.26.20
whitenoise==6.8.2
playwright
pdfplumber==0.10.0
PyPDF2==3.0.1
python-dateutil==2.8.2
